Modify counting for primary + full tags
The previous instance took the tags list (comma separated) and just made that
the unit of tracking. This version splits it into a primary project, which it
counts while also tracking the full tagged list. Modified the printout to
first print the primary counts, keeping it to 100% coverage and single
counting.