Fix bug on TimeRecord Comparison
- fix bug
- modify time comparison for inRange to include date and tod

The prior version did not include the date, but a TimeRecord tracked
later by a different timezone would create a time on 5.10.23 @ 9:00 but
before conversion it would show the date ahead relative to a prior time
and implicitliy organize it later. Or something like this. I'm not
clear. The solution here is to include the date extraction as a part of
the comparison which correctly orders.

This commit also fixes the comparison for the inRange function to
include only date and time but not timezone and also uses <= and >= to
capture 0:00 properly.

Signed-off-by: Nathan Dautenhahn <ndd@rice.edu>